+++ workspaces/orchestrator/plugins/orchestrator/src/components/WorkflowPage/RunButton.tsx
@@ -14,7 +14,6 @@
  * limitations under the License.
  */
 
-import React from 'react';
 import { useNavigate } from 'react-router-dom';
 
 import { useRouteRef, useRouteRefParams } from '@backstage/core-plugin-api';
@@ -32,12 +31,22 @@ import {
 import { usePermissionArrayDecision } from '../../hooks/usePermissionArray';
 import { executeWorkflowRouteRef, workflowRouteRef } from '../../routes';
 
-export const RunButton = ({ isAvailable }: { isAvailable?: boolean }) => {
+export const RunButton = ({
+  isAvailable,
+  entityRef,
+}: {
+  isAvailable?: boolean;
+  entityRef?: string;
+}) => {
   const { workflowId } = useRouteRefParams(workflowRouteRef);
   const navigate = useNavigate();
   const executeWorkflowLink = useRouteRef(executeWorkflowRouteRef);
   const handleExecute = () => {
-    navigate(executeWorkflowLink({ workflowId }));
+    navigate(
+      entityRef
+        ? `${executeWorkflowLink({ workflowId })}?targetEntity=${entityRef}`
+        : executeWorkflowLink({ workflowId }),
+    );
   };
 
   const { loading: loadingPermission, allowed: canRun } =