+++ workspaces/orchestrator/plugins/orchestrator-form-api/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @red-hat-developer-hub/backstage-plugin-orchestrator-form-api
 
+## 2.0.2
+
+### Patch Changes
+
+- Updated dependencies [26e602a]
+  - @red-hat-developer-hub/backstage-plugin-orchestrator-common@3.0.2
+
+## 2.0.1
+
+### Patch Changes
+
+- Updated dependencies [32e0a44]
+  - @red-hat-developer-hub/backstage-plugin-orchestrator-common@3.0.1
+
 ## 2.0.0
 
 ### Major Changes